requirements.txt
flask
requests
python-dotenv
openai
pytest